refactor: move contract issues to Github class
