+++ packages/github/tests/contract/list_issues.test.mjs
@@ -1,19 +1,13 @@
+import { MAX_ISSUES_PER_PAGE } from "github_issues/src/constants.mjs";
 import { setupServer } from "msw/node";
+import { buildIssues } from "testing/src/builders/issues.mjs";
+import { listIssuesFailureHandler, listIssuesHandler } from "testing/src/interceptors/issues_handler.mjs";
 import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest";
-import { buildGithubClient, buildGithubContext, buildGithubCore } from "../../../testing/src/builders/github_core.mjs";
-import { buildIssues } from "../../../testing/src/builders/issues.mjs";
-import { listIssuesFailureHandler, listIssuesHandler } from "../../../testing/src/interceptors/issues_handler.mjs";
-import { MAX_ISSUES_PER_PAGE } from "../../src/constants.mjs";
-import { listIssues } from "../../src/issues.mjs";
-
-const org = "aws-powertools";
-const repo = "powertools-lambda-python";
+import { Github } from "../../src/Github.mjs";
 
 describe("list issues contract", () => {
+	process.env.GITHUB_REPOSITORY = "test-org/test-repo";
 	const server = setupServer();
-	const github = buildGithubClient({ token: process.env.GITHUB_TOKEN });
-	const context = buildGithubContext({ org, repo });
-	const core = buildGithubCore();
 
 	beforeAll(() => server.listen({ onUnhandledRequest: "error" }));
 
@@ -23,15 +17,12 @@ describe("list issues contract", () => {
 
 	it("should list issues (default parameters)", async () => {
 		// GIVEN
+		const github = new Github();
 		const existingIssues = buildIssues({ max: 5 });
-		server.use(...listIssuesHandler({ data: existingIssues, org, repo }));
+		server.use(...listIssuesHandler({ data: existingIssues, org: github.owner, repo: github.repo }));
 
 		// WHEN
-		const ret = await listIssues({
-			github,
-			context,
-			core,
-		});
+		const ret = await github.listIssues();
 
 		// THEN
 		expect(ret).toStrictEqual(existingIssues);
@@ -39,34 +30,27 @@ describe("list issues contract", () => {
 
 	it("should filter out pull requests from results", async () => {
 		// GIVEN
+		const github = new Github();
 		const realIssues = buildIssues({ max: 2 });
 		const prAsIssues = buildIssues({ max: 2, isPr: true });
-		server.use(...listIssuesHandler({ data: [...realIssues, ...prAsIssues], org, repo }));
+		server.use(...listIssuesHandler({ data: [...realIssues, ...prAsIssues], org: github.owner, repo: github.repo }));
 
 		// WHEN
-		const ret = await listIssues({
-			github,
-			context,
-			core,
-		});
+		const ret = await github.listIssues();
 
 		// THEN
 		expect(ret).toStrictEqual(realIssues);
 	});
 
 	it("should exclude results with certain labels", async () => {
+		const github = new Github();
 		const BLOCKED_LABELS = "do-not-merge";
 
 		const existingIssues = buildIssues({ max: 5, labels: [BLOCKED_LABELS] });
-		server.use(...listIssuesHandler({ data: existingIssues, org, repo }));
+		server.use(...listIssuesHandler({ data: existingIssues, org: github.owner, repo: github.repo }));
 
 		// WHEN
-		const ret = await listIssues({
-			github,
-			context,
-			core,
-			excludeLabels: BLOCKED_LABELS,
-		});
+		const ret = await github.listIssues({ excludeLabels: BLOCKED_LABELS });
 
 		// THEN
 		expect(ret.length).toBe(0);
@@ -74,18 +58,14 @@ describe("list issues contract", () => {
 
 	it("should limit the number of issues returned", async () => {
 		// GIVEN
+		const github = new Github();
 		const maxIssuesToReturn = 1;
 		const existingIssues = buildIssues({ max: 2 });
 
-		server.use(...listIssuesHandler({ data: existingIssues, org, repo }));
+		server.use(...listIssuesHandler({ data: existingIssues, org: github.owner, repo: github.repo }));
 
 		// WHEN
-		const ret = await listIssues({
-			github,
-			context,
-			core,
-			limit: maxIssuesToReturn,
-		});
+		const ret = await github.listIssues({ limit: maxIssuesToReturn });
 
 		// THEN
 		expect(ret.length).toBe(maxIssuesToReturn);
@@ -93,18 +73,14 @@ describe("list issues contract", () => {
 
 	it("should paginate to list all available issues when the limit is higher", async () => {
 		// GIVEN
+		const github = new Github();
 		const totalCount = MAX_ISSUES_PER_PAGE + 1;
 
 		const existingIssues = buildIssues({ max: totalCount });
-		server.use(...listIssuesHandler({ data: existingIssues, org, repo }));
+		server.use(...listIssuesHandler({ data: existingIssues, org: github.owner, repo: github.repo }));
 
 		// WHEN
-		const ret = await listIssues({
-			github,
-			context,
-			core,
-			limit: totalCount,
-		});
+		const ret = await github.listIssues({ limit: totalCount });
 
 		// THEN
 		expect(ret.length).toBe(totalCount);
@@ -112,17 +88,12 @@ describe("list issues contract", () => {
 
 	it("should throw error when GitHub API call fails (http 500)", async () => {
 		// GIVEN
+		const github = new Github();
 		const err = "Unable to process request at this time";
-		server.use(...listIssuesFailureHandler({ org, repo, err }));
+		server.use(...listIssuesFailureHandler({ org: github.owner, repo: github.repo, err }));
 
 		// WHEN
 		// THEN
-		const ret = await expect(
-			listIssues({
-				github,
-				context,
-				core,
-			}),
-		).rejects.toThrowError(err);
+		const ret = await expect(github.listIssues()).rejects.toThrowError(err);
 	});
 });
