--- packages/github_pull_requests/tests/contract/list_pull_request.test.mjs/pull_requests.test.mjs
@@ -1,116 +1,0 @@
-import { setupServer } from "msw/node";
-import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest";
-import {
-	buildGithubClient,
-	buildGithubContext,
-	buildGithubCore,
-} from "../../../../testing/src/builders/github_core.mjs";
-import { buildPullRequests } from "../../../../testing/src/builders/pull_requests.mjs";
-import {
-	listPullRequestsFailureHandler,
-	listPullRequestsHandler,
-} from "../../../../testing/src/interceptors/pull_requests_handler.mjs";
-import { MAX_PULL_REQUESTS_PER_PAGE } from "../../../src/constants.mjs";
-import { listPullRequests } from "../../../src/pull_requests.mjs";
-
-const org = "aws-powertools";
-const repo = "powertools-lambda-python";
-
-describe("list pull requests contract", () => {
-	const server = setupServer();
-	const github = buildGithubClient({ token: process.env.GITHUB_TOKEN });
-	const context = buildGithubContext({ org, repo });
-	const core = buildGithubCore();
-
-	beforeAll(() => server.listen({ onUnhandledRequest: "error" }));
-
-	afterAll(() => server.close());
-
-	afterEach(() => server.resetHandlers());
-
-	it("should list pull requests (default parameters)", async () => {
-		// GIVEN
-		const data = buildPullRequests({ max: 5 });
-		server.use(...listPullRequestsHandler({ data, org, repo }));
-
-		// WHEN
-		const ret = await listPullRequests({
-			github,
-			context,
-			core,
-		});
-
-		// THEN
-		expect(ret).toStrictEqual(data);
-	});
-
-	it("should exclude results with certain labels", async () => {
-		// GIVEN
-		const BLOCKED_LABELS = "do-not-merge";
-		const data = buildPullRequests({ max: 5, labels: [BLOCKED_LABELS] });
-		server.use(...listPullRequestsHandler({ data, org, repo }));
-
-		// WHEN
-		const ret = await listPullRequests({
-			github,
-			context,
-			core,
-			excludeLabels: BLOCKED_LABELS,
-		});
-
-		// THEN
-		expect(ret.length).toBe(0);
-	});
-
-	it("should limit the number of pull requests returned", async () => {
-		// GIVEN
-		const data = buildPullRequests({ max: 2 });
-		server.use(...listPullRequestsHandler({ data, org, repo }));
-
-		// WHEN
-		const ret = await listPullRequests({
-			github,
-			context,
-			core,
-			limit: 1,
-		});
-
-		// THEN
-		expect(ret.length).toBe(1);
-	});
-
-	it("should paginate to list all available pull requests when the limit is higher", async () => {
-		// GIVEN
-		const totalPRs = MAX_PULL_REQUESTS_PER_PAGE + 1;
-
-		const data = buildPullRequests({ max: totalPRs });
-		server.use(...listPullRequestsHandler({ data, org, repo }));
-
-		// WHEN
-		const ret = await listPullRequests({
-			github,
-			context,
-			core,
-			limit: totalPRs,
-		});
-
-		// THEN
-		expect(ret.length).toBe(totalPRs);
-	});
-
-	it("should throw error when GitHub API call fails (http 500)", async () => {
-		// GIVEN
-		const err = "Unable to process request at this time";
-		server.use(...listPullRequestsFailureHandler({ org, repo, err }));
-
-		// WHEN
-		// THEN
-		const ret = await expect(
-			listPullRequests({
-				github,
-				context,
-				core,
-			}),
-		).rejects.toThrowError(err);
-	});
-});
